+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.7.3] - 2024-02-12
 ### Stable Release
 - Add User's export feature
+- Add User' Type support API
 
 ## [2.7.2] - 2024-02-11
 ### Stable Release